gistrees-server/requirements.txt
Flask
gunicorn
psycopg2-binary
Pillow
flask-cors